LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-center-vrl-htb-expected.txt


FAIL .target 1 assert_equals:
<div class="container">
  <div class="anchor"></div>
  <div class="target" data-expected-width="100" data-offset-x="0"></div>
</div>
width expected 100 but got 50
FAIL .target 2 assert_equals:
<div class="container">
  <div class="anchor"></div>
  <div class="target" style="max-width: 60px;" data-expected-width="60" data-offset-x="35"></div>
</div>
width expected 60 but got 50
PASS .target 3
FAIL .target 4 assert_equals:
<div class="container">
  <div class="anchor"></div>
  <div class="target" style="right: 20px;" data-expected-width="80" data-offset-x="20"></div>
</div>
offsetLeft expected 20 but got 0
PASS .target 5
FAIL .target 6 assert_equals:
<div class="container">
  <div class="anchor"></div>
  <div class="target" style="max-width: 100px; right: -20px;" data-expected-width="100" data-offset-x="15"></div>
</div>
offsetLeft expected 15 but got 20
PASS .target 7
FAIL .target 8 assert_equals:
<div class="container">
  <div class="anchor"></div>
  <div class="target" style="left: 10px; right: 20px;" data-expected-width="70" data-offset-x="30"></div>
</div>
offsetLeft expected 30 but got 10
PASS .target 9
PASS .target 10
